+++ Modules/Module.psm1
@@ -1,6 +1,7 @@
 <#
-
 Prevents autoloading subfolders
+
+If name = Modules.psm1 it wont auto-load any directory
 
 Remove or rename this file and Modules will be auto-loaded when placed in a $env:PSModulePath